requirements.txt
# AI Frameworks
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.8.0+cu126
torchaudio==2.8.0+cu126
torchvision
accelerate
# Flash Attention
https://github.com/Dao-AILab/flash-attention/releases/download/v2.8.3/flash_attn-2.8.3+cu12torch2.8cxx11abiFALSE-cp310-cp310-linux_x86_64.whl

# AI Models and Tools
diffusers
transformers
qwen-tts

# Server
fastapi
uvicorn[standard]
pillow

